FEAT: Functionality of additional recipe pages
Features added: Description, Materials, Ingredients, Steps.

Previously a blank page, this page now informs a reader how to cook
the meal.  This includes a description for the user of the meal,
the materials and ingredients they will need, and the steps needed
to take to cook the meal.

Formatting used the emphasis of italics and bold font to highlight
ingredients to make it visually easier to read.

Additional functionality added through adding a redirect to home page
link at the bottom.

Adjusted coding format for lemony-chicken-spaghetti recipe to match
other recipes.