FIX: Basic CSS file added to project
Prior to this no CSS had been used on the project. As a part of
learning the basics of CSS, I was to implement the basic concepts into
this project.  These concepts include styling various properties
such as color, background color, typography properties, image
properties, etc.  The Cascade of CSS is taken into account when
assigning div classes and ids with their respective properties.
In addition all properties are assigned externally as opposed to
internally or inline for better practice.